+++ src/mieic/iart/SignLangNN/neuralnetwork/Edge.java
@@ -12,16 +12,16 @@ public class Edge {
     private Node sourceNode;
     private Node destinationNode;
 
-    private float weight;
-    private float savedValue;
-    private float error;
+    private double weight;
+    private double savedValue;
+    private double error;
 
     public Edge(Node source, Node destination) {
         this.sourceNode = source;
         this.destinationNode = destination;
 
         Random r = new Random();
-        weight = r.nextFloat() * 2 - 1;
+        weight = r.nextDouble() * 2 - 1;
         error = 0;
     }
 
@@ -41,41 +41,41 @@ public class Edge {
         this.destinationNode = destinationNode;
     }
 
-    public float getWeight() {
+    public double getWeight() {
         return weight;
     }
 
-    public void setWeight(float weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
         Log.log("Weight set to " + weight);
     }
 
-    public void inflowValue(float val) {
+    public void inflowValue(double val) {
         savedValue = val;
         Log.log("Value received from previous node: " + val);
     }
 
-    public float outflowValue() {
-        float outflowValue = savedValue * weight;
+    public double outflowValue() {
+        double outflowValue = savedValue * weight;
         Log.log("Sent value to next node: " + outflowValue);
         return outflowValue;
     }
 
     // training functions
 
-    public void setError(float error) {
+    public void setError(double error) {
         this.error = error;
     }
 
-    public float backPropagateError() {
-        float weightedError = error * weight;
+    public double backPropagateError() {
+        double weightedError = error * weight;
         Log.log("Feeding-backward the error. Value: " + weightedError);
         return weightedError;
     }
 
     public void updateWeight() {
-        float oldWeight = weight;
-        weight = weight + NeuralNetwork.LEARNING_RATE * destinationNode.getError() * (float) (Math.exp(sourceNode.getOutput())
+        double oldWeight = weight;
+        weight = weight + NeuralNetwork.LEARNING_RATE * destinationNode.getError() * (double) (Math.exp(sourceNode.getOutput())
                 / Math.pow(Math.exp(sourceNode.getOutput()) + 1, 2.0));
 
         Log.log("Weight changed from " + oldWeight + " to " + weight);
